refactor(giscus-comments): move `GiscusComments` component to shared components folder
